tests/wpt/tests/navigation-api/commit-behavior/redirect-throws.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
promise_test(async t => {
  navigation.onnavigate = t.step_func(e => {
    assert_throws_dom("InvalidStateError", () => e.redirect("#"));
  });
  await navigation.navigate("#").finished;
}, "redirect() before intercept()");

promise_test(async t => {
  navigation.onnavigate = t.step_func(e => {
    e.intercept({ handler: t.step_func(() => {
      assert_throws_dom("InvalidStateError", () => e.redirect("#"));
    }) });
  });
  await navigation.navigate("#").finished;
}, "redirect() without commit behavior specified");

promise_test(async t => {
  navigation.onnavigate = t.step_func(e => {
    e.intercept({
      handler: t.step_func(() => {
        assert_throws_dom("InvalidStateError", () => e.redirect("#"));
      }),
      commit: "immediate"
    });
  });
  await navigation.navigate("#").finished;
}, "redirect() with { commit: immediate }");

promise_test(async t => {
  navigation.onnavigate = t.step_func(e => {
    e.intercept({ commit: "after-transition" });
    assert_throws_dom("InvalidStateError", () => e.redirect("#"));
  });
  await navigation.navigate("#").finished;
}, "redirect() during event dispatch");

promise_test(async t => {
  let navigate_event;
  navigation.onnavigate = t.step_func(e => {
    e.intercept({ commit: "after-transition" });
    navigate_event = e;
  });
  await navigation.navigate("#").finished;
  assert_throws_dom("InvalidStateError", () => navigate_event.redirect("#"));
}, "redirect() after finish");

promise_test(async t => {
  navigation.onnavigate = t.step_func(e => {
    e.intercept({
      handler: t.step_func(() => {
        e.commit();
        assert_throws_dom("InvalidStateError", () => e.redirect("#"));
      }),
      commit: "after-transition"
    });
  });
  await navigation.navigate("#").finished;
}, "redirect() after commit()");

promise_test(async t => {
  // We need to grab an NavigationDestination to construct the event.
  navigation.onnavigate = t.step_func(e => {
    const event = new NavigateEvent("navigate", {
      destination: e.destination,
      signal: (new AbortController()).signal
    });

    assert_throws_dom("SecurityError", () => event.redirect("#"));
  });
  await navigation.navigate("#").finished;
}, "redirect() on synthetic NavigateEvent");

promise_test(async t => {
  let i = document.createElement("iframe");
  i.src = "about:blank";
  document.body.appendChild(i);
  i.contentWindow.navigation.onnavigate = t.step_func(e => {
    e.intercept({
      handler: t.step_func(() => {
        let iframe_constructor = i.contentWindow.DOMException;
        i.remove();
        assert_throws_dom("InvalidStateError", iframe_constructor, () => e.redirect("#"));
      }),
      commit: "after-transition"
    });
  });
  i.contentWindow.navigation.navigate("#");
}, "redirect() in detached iframe");

promise_test(async t => {
  navigation.onnavigate = t.step_func(e => {
    e.intercept({ handler: t.step_func(() => {
      assert_throws_dom("SyntaxError", () => e.redirect("https://example.com\u0000mozilla.org"));
    }),
    commit: "after-transition" });
  });
  await navigation.navigate("#").finished;
}, "redirect() to invalid url");

promise_test(async t => {
  navigation.onnavigate = t.step_func(e => {
    e.intercept({ handler: t.step_func(() => {
      assert_throws_dom("SecurityError", () => e.redirect("https://example.com"));
    }),
    commit: "after-transition" });
  });
  await navigation.navigate("#").finished;
}, "redirect() to cross-origin url");

promise_test(async t => {
  navigation.onnavigate = t.step_func(e => {
    e.intercept({ handler: t.step_func(() => {
      assert_throws_dom("InvalidStateError", () => e.redirect("#"));
    }),
    commit: "after-transition" });
  });
  await navigation.reload().finished;
}, "redirect() reload");

promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  await navigation.navigate("#forward").finished;

  navigation.onnavigate = t.step_func(e => {
    e.intercept({ handler: t.step_func(() => {
      assert_throws_dom("InvalidStateError", () => e.redirect("#"));
    }),
    commit: "after-transition" });
  });
  await navigation.back().finished;
}, "redirect() traverse");
</script>
</body>
